renamed files to avoid conflicts on case-insensitive filesystems
Apparently MacOS is case-insensitive by default. This closes #21.
