+++ Xkw/lodemo.c
