+++ dominos/Dominos.h
@@ -25,7 +25,7 @@
 
 #ifndef _XtDominos_h
 #define _XtDominos_h
-#include "dominos.h"
+#include "domino.h"
 #include <Xkw/Xkw.h>
 
 /* define exposed functions */
