+++ dominos/connect.c
@@ -23,5 +23,5 @@
  * Author:  Keith Packard, Network Computing Devices
  */
 
-#include "dominos.h"
+#include "domino.h"
